Fix the tests
