+++ app/tests/Entity/PlayerTest.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Tests\App\Entity;
+
+use PHPUnit\Framework\TestCase;
+use App\Entity\Player;
+
+class PlayerTest extends TestCase {
+  public function testAddTurn() {
+    $player = new Player(
+      'randomtoken',
+      array(
+        array(1, 2, 3),
+        array(4, 0, 5),
+        array(7, 8, 6)
+      )
+    );
+
+    for ($i = 1; $i <= 23; $i++) {
+      $player->addTurn();
+    }
+
+    $this->assertEquals(23, $player->getTurn());
+  }
+}
